Refactored UserToolTip
This version has been re-written to be more concise, easier to read and improved modularity.

Added: ActionButton which is a component for the action buttons underneath the name and jobtitle.
Added: UserCardActions which contains the ActionButtons.
Added: Classes for all jsx elements to tidy up JSX and remove inline styles.
Added: UserDetailsWrapper to surround UserDetails and UserCardActions.
Added: UserInfoContainer to surround UserContactDetails and UserCardDepartments, also fixes the padding issue with the info before it extends in height showing its padding as padding has been moved into a nested div and can now be hidden.

Changed: Entire naming convention of component and some preexisting components from the UserToolTip variant/version.